+++ skins/skypotato/uibits.class.php
@@ -105,6 +105,10 @@ class uibits {
 		return $h;
 	}
 
+	public static function browse_artistHeader($id, $name) {
+		return '';
+	}
+
 	public static function noAlbumsHaeder() {
 		print '<div class="playlistrow2" style="padding-left:64px">'.language::gettext("label_noalbums").'</div>';
 	}